Playable test added.
Adds a new test scene "Play". This scene should be playable in a minimal
sense. Obstacles spawn in the distance (10s ahead of the player) and
despawn as they pass the player (5s behind the player). Obstacles are
stationary only the player and camera are moving (relativity is
confusing). The player reacts as before to collisions. Spawns are random
(using Godot's rather than C#/Microsoft's PRNG implementation) with a
maximum of 8 obstacles spawning in every half second.

This also does away with GridMaps and instead enforces movement
constraints based on vectors alone. If this is an issue perhaps it could
be brought.