Update to more straight-forward level format.
Changes the format of Level JSON files to be more straight forward (in
my opinion) this also provides support describing entities
configured to translate or scale (this isn't implemented yet).

Additionally, the basic entities are defined under
scenes/prefab/entities (fields and letterboxes will need more work).

Finally, there are some small modifications to the player and level
loading scenes. The level loading modifications are to support the new
format. For the player I've just changed the base color so it's easier
to see.