Add level lead in/out behavior.
This adds a 3 second lead in and out time at the beginning and end of a
stage. Perhaps a countdown would be shown during the lead in period.
During the lead out perhaps the camera could freely pan around the
player to show the level or something.

After the lead out time is complete the game transitions back the the
title menu but in the future it probably makes more sense to transition
to the next level, a status screen, or some other such thing.