Add basic game save handling.
Changes the name of the Utility object to GameMain. This object now
holds the global play data. When the game starts, GameMain is
automatically loaded and inserted into the tree. When ready, this
triggers loading the save file. When the GameMain object exits the tree
(at shutdown) the save data is written back to the disk. As long as the
data file stays small and simple this shouldn't be an issue.