Update PressAnyKey.
Moves PressAnyKey.tscn to scenes/prefab because it has broader use
outside of the Menu.

Adds a toggle to PressAnyKey to control fading/blinking animation (only
read at _Ready currently).

Changes behavior on _UnhandledInput so that button release events and
repeat (echo) events are not considered. This prevents the Pressed
signal from firing multiple times in a row.