Re #1: Implement remaining root scene transitions.
Changes and fixes to implement the remaining root scene transitions
(Play, GameComplete, GameOver). This adds basic implementations of the
Game Complete/Game Over scene. To enable sharing of behavior between the
two possible end scenes a GameEndHandler node is also added.

Logic is added to MenuRoot to select the correct menu scene depending on
whether or not the previous scene was an ending (which transitions
directly to Title B). To support this logic is added to the GameMain
object to track the current and previous RootScene.

PlayRoot is implemented as a stub that immediately transitions (at
random to an ending).