Add Field behavior.
Adds a script for the Field entity. Fields interact with objects that
have the interface IVelocityModifiable. When an object with that
interface enters a field IVelocityModifiable.ModifyVelocity is called to
set all movement (including rotation) half speed. When the object exits
the rates are restored to one.

Adds IVelocityModifiable interface. Objects implementing this interface
are affected by Fields and are responsible for ensuring that their
movement speed is actually reduced as a result.