Add new Player entity.
Adds a new version of the Player entity. A lot of functionality is
missing (e.g., movement) and some parts of the scene are temporary to
support testing. Currently rotation works as per documentation. The
player can rotate infinitely in either the clockwise or
counter-clockwise direction. Rotations are 90 degree (PI/2 radian)
increments. There is a 0.5s delay between each completed rotation.