Add movement back to Player entity.
The Player entity now has its X and Y axis movement behavior restored.
To support this a BoundedKinematicBody class is added which contains
shared behavior for entities with bounded (to the game grid) X and Y
axis movement. Right now this is just the Player and the Cube.

The Cube entity is updated as appropriate for the new
BoundedKinematicBody base class.