Re #2: Add Camera translation effect.
This adds a translation effect to the FollowCamera. When the target node
(probably the Player) moves along the X or Y axis the FollowCamera will
interpolate to follow it. This takes 0.5s currently. I think the effect
looks ok relative to other objects in the game world. Without reference
objects and Godot's default procedural horizon it looks rather odd.