Add level transition behavior.
This adds new behavior to transition from one level to another (provided
an override is not set) during play. Levels are transitioned to in order
as provided in the Levels array. When the player completes the final
level the game will transition the GameComplete screen. If the game
has already been completed then Continue will replay the final level and
result in a GameComplete when finished successfully. The PlayRoot scene
is modified accordingly to test this behavior.

Adds a name field to BasicLevel. Currently this isn't necessary but it
might help to further identify levels. Might remove it later.

Adds more integer clamps to Utility.

Adds a slightly modified duplicate test level for testing transitions.