Implement pause menu.
This makes various changes to support the pause menu. Most obviously, it
adds the PauseScreen and PauseMenu scenes. It also makes changes to
PlayRoot to handle requests for restarting the game or resuming from the
menu. This also alters level restart behavior so that the player is
always initialized (0 shields, 2 health, rotation 0) at the start of a
level.

This also changes the Settings object slightly so that it can interact
properly with both its possible parents.