Add flag field to PlayerData.
This adds a 16-bit field to the PlayerData object containing an
unspecified set of bit flags. There was already a 16-bit wide pad in the
file specification so I've converted that into space for flags. This is
to separate the idea of an "initialized" player from a player who hasn't
gone through the tutorial level yet.

File format documentation will be updated at a later time.