Further tutorial level infrastructure.
The tutorial level is now broken out into a separate property of
PlayRoot. It is only loaded if the string is actually set (currently
it's set to res://scenes/level/Tutorial.tscn). When the tutorial level
is completed the player is flagged and proceeds directly into the
regular level list. On repeat plays the flag is checked and play
proceeds directly into the regular level list. There isn't an in-game
method to clear the tutorial flag but it can be cleared by toggling
"Initialize Player Flags On Start" in the inspector for PlayRoot. This
will create a totally blank PlayerData each time PlayRoot is loaded. The
file can also be initialized by deleting it entirely (on Linux systems
the file is located at
$HOME/.local/share/godot/app_userdata/obscure_game/save.bin).

This also adds a "Tutorial" level (currently empty) and sets the names
of the currently existing test levels. Levels are now identified by
their LevelName property and, accordingly, LevelName has been promoted
from BasicLevel to Level.

Finally, a bug regarding GameMain.InitializePlayerData() is fixed. It
wasn't clearing flags when told to. Now it works as intended.