Re #6: Shrink Player width and depth.
Previously, the Player was exactly 1 unit (1 meter) wide and 1 unit
deep. This changes the Player to be 0.75 units wide and 0.75 units deep.
Corresponding changes are made to the Player entity's collision box.
This prevents marginal collisions when a 1 unit player passes through a
1 unit volume.

This change has an effect on the visuals of the game. If necessary in
the future we can retain the smaller collision box and change the
Player's visual representation to something larger.

Resolves #6